Save during training, load during test
When Saver.save() is called, it saves 2 files and overwrites 1 file:
- checkpoint: binary that keeps track of latest checkpoints
- <epoch_ckpt>: weights for that epoch
- <epoch_ckpt>.meta: graph

.meta file defines graph. Not sure what else it defines because graph
should be same every time?

Test plan:
1) First, make some checkpoints (break out of minibatch loop to do this quickly)
- python tasks/image-sent/main.py --arch basic_cnn -obj sent_biclass --mode train --save_every_epoch 1

2) Load latest checkpoint
- python tasks/image-sent/main.py --arch basic_cnn -obj sent_biclass --mode test --ckpt_dir 2016-12-22___13-30-18
- print some weights: sess.run(tf.trainable_variables()[0])

3) Load a specific checkpoint - test 1: see that loading latest matches 2)
- python tasks/image-sent/main.py --arch basic_cnn -obj sent_biclass --mode test --ckpt_dir 2016-12-22___13-30-18 --load_epoch 9
- check weights match

4) Load a specific checkpoint - test 2: see that weights are different from 3)
- python tasks/image-sent/main.py --arch basic_cnn -obj sent_biclass --mode test --ckpt_dir 2016-12-22___13-30-18 --load_epoch 3
- check weights are different

Resource: http://stackoverflow.com/questions/33759623/tensorflow-how-to-restore-a-previously-saved-model-python

Resource for loading:

+++ visual-tensorflow/bash/cleanup_checkpoints.sh
@@ -8,8 +8,10 @@ for task_dir in */ ; do
     cd $task_dir/checkpoints
     for ckpt_dir in */ ; do
         if [ -d ${ckpt_dir} ]; then        # will not run if no directories are available
-            if [ $(ls $ckpt_dir) == "params.json" ]; then
-                rm -rf $ckpt_dir
+            if [ $(ls $ckpt_dir | wc -l) == 1 ]; then
+                if [ $(ls $ckpt_dir) == "params.json" ]; then
+                    rm -rf $ckpt_dir
+                fi
             fi
         fi
     done
